add hrefs to application page when double click application box
