+++ frontend/pages/homepage.html
@@ -72,17 +72,17 @@
       <div class="status-columns">
         <div class="status-column" id="interested">
           <h3>Interested</h3>
-          <div class="application-box" draggable="true" id="amazon">Amazon<br>• Software Engineer</div>
+          <div class="application-box" draggable="true" id="amazon" ondblclick="window.location.href='applicationInfo.html'">Amazon<br>• Software Engineer</div>
 
         </div>
         <div class="status-column" id="applied">
           <h3>Applied</h3>
-          <div class="application-box" draggable="true" id="dell">Dell<br>• Software Engineer</div>
+          <div class="application-box" draggable="true" id="dell" ondblclick="window.location.href='applicationInfo.html'">Dell<br>• Software Engineer</div>
 
         </div>
         <div class="status-column" id="interviewing">
           <h3>Interviewing</h3>
-          <div class="application-box" draggable="true" id="google">Google<br>• Software Developer</div>
+          <div class="application-box" draggable="true" id="google" ondblclick="window.location.href='applicationInfo.html'">Google<br>• Software Developer</div>
 
         </div>
         <div class="status-column" id="offer">
@@ -101,7 +101,7 @@
       </div>
     </div>
   </div>
-  <script src="../scripts/script.js"></script>
+  <script type="module" src="../scripts/script.js"></script>
 
   <!-- Footer Section -->
   <div class="footer-section">
